Move profiling into dev_scripts/profile.py
Also add makefile recipe
