+++ tests/test_bgzip.py
@@ -2,11 +2,9 @@
 import io
 import os
 import sys
-import time
 import gzip
 import unittest
 from random import randint
-from contextlib import AbstractContextManager
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
@@ -108,6 +106,7 @@ class TestBGZipWriter(unittest.TestCase):
                     reinflated_data += d
                 else:
                     break
+
         self.assertEqual(inflated_data, reinflated_data)
         self.assertTrue(fh_out.getvalue().endswith(bgzip.bgzip_eof))
 
@@ -115,73 +114,6 @@ class TestBGZipWriter(unittest.TestCase):
         fh = io.BytesIO()
         with bgzip.BGZipWriter(fh):
             fh.write(b"")
-
-class TestProfileBGZip(unittest.TestCase):
-    def test_profile_read(self):
-        print()
-        with open("tests/fixtures/partial.vcf.gz", "rb") as raw:
-            with profile("gzip read"):
-                with gzip.GzipFile(fileobj=raw) as fh:
-                    fh.read()
-
-        for num_threads in range(1, 1 + bgzip.available_cores):
-            with profile(f"BGZipReader (num_threads={num_threads})"):
-                with open("tests/fixtures/partial.vcf.gz", "rb") as raw:
-                    with bgzip.BGZipReader(raw, num_threads=num_threads) as reader:
-                        while True:
-                            data = reader.read(randint(1024 * 1024 * 1, 1024 * 1024 * 10))
-                            if not data:
-                                break
-
-    def test_profile_write(self):
-        print()
-        with open("tests/fixtures/partial.vcf.gz", "rb") as raw:
-            with gzip.GzipFile(fileobj=raw) as fh:
-                inflated_data = fh.read()
-
-        with profile("gzip write"):
-            with gzip.GzipFile(fileobj=io.BytesIO(), mode="w") as fh:
-                fh.write(inflated_data)
-
-        for num_threads in range(1, 1 + bgzip.available_cores):
-            with profile(f"bgzip write (num_threads={num_threads})"):
-                with bgzip.BGZipWriter(io.BytesIO(), num_threads=num_threads) as writer:
-                    n = 987345
-                    writer.write(inflated_data[:n])
-                    writer.write(inflated_data[n:])
-        print()
-
-class profile(AbstractContextManager):
-    """
-    Profile methods or code blocks with decorators or contexts, respectively.
-
-    @profile("profiling my method")
-    def my_method(*args, **kwargs):
-       ...
-
-    with profile("profiling my block"):
-       ...
-    """
-    def __init__(self, name="default"):
-        self.name = name
-
-    def __enter__(self, *args, **kwargs):
-        self.start = time.time()
-
-    def __exit__(self, *args, **kwargs):
-        self._print(time.time() - self.start)
-
-    def _print(self, duration):
-        print(f"{self.name} took {duration} seconds")
-
-    def __call__(self, meth):
-        def wrapper(*args, **kwargs):
-            start = time.time()
-            res = meth(*args, **kwargs)
-            dur = time.time() - start
-            self._print(dur)
-            return res
-        return wrapper
 
 if __name__ == '__main__':
     unittest.main()
